Add segment_sum functions in pytorch_utils.py
